use new queue, remove old implementation
using actual queue instead of fake
